perf(site): split landing into server/client islands + Suspense for session
The homepage was a single "use client" tree where the page-level Promise.all
awaited a session lookup (via headers()) on every render, opting the entire
route out of static rendering and forcing the slowest data source to gate
first paint.

Refactor:
- New components/site/landing-static-sections.tsx (server component) with
  StackSection + ContentSection. These render static content; their only
  client dependency is ScrollReveal, which can be a client child of a server
  component.
- components/site/landing-sections.tsx is slimmed to two client islands —
  HeroSection and InstallSection — both exported. Drop the LandingClient
  wrapper, the dead ProjectsSection, the dead ReposSection, and unused
  imports (ContentSummary, GitHubRepo, ContentCard, formatDate, stack).
- app/(site)/page.tsx composes directly:
  - Drop unused getPinnedProjects() and getRecentRepos() — their results
    were never rendered (LandingClient ignored projects/repos props).
  - Wrap the session-dependent hero in <Suspense fallback={<HeroSection
    userName={null} />}> so anonymous traffic streams the static-rendered
    fallback immediately and the personalized greeting swaps in once
    PersonalizedHero (server async) resolves.

Net effect on the homepage:
- Two fewer cache lookups + GitHub fetch on every render.
- The static landing tree (Stack/Install/Content) is now eligible for
  the cacheComponents static partition; only PersonalizedHero is dynamic.
- Bundle: removes ContentCard/formatDate/stack from the "use client"
  module, shrinking what hydrates with the page.

Trade-off: for logged-in visitors, a brief fallback→personalized swap when
the session promise resolves. Acceptable: session typically resolves under
100 ms, well before the input animation completes.
