+++ apps/chat/app/(site)/page.tsx
@@ -1,9 +1,13 @@
 import { headers } from "next/headers";
-import { LandingClient } from "@/components/site/landing-sections";
+import { Suspense } from "react";
+import { HeroSection, InstallSection } from "@/components/site/landing-sections";
+import {
+  ContentSection,
+  StackSection,
+} from "@/components/site/landing-static-sections";
 import { getSafeSession } from "@/lib/auth";
 import { config } from "@/lib/config";
-import { getLatest, getPinnedProjects } from "@/lib/content";
-import { getRecentRepos } from "@/lib/github";
+import { getLatest } from "@/lib/content";
 
 const jsonLd = {
   "@context": "https://schema.org",
@@ -24,12 +28,9 @@ const jsonLd = {
 };
 
 export default async function Home() {
-  const [projects, writing, notes, repos, session] = await Promise.all([
-    getPinnedProjects(3),
+  const [writing, notes] = await Promise.all([
     getLatest("writing", 3),
     getLatest("notes", 3),
-    getRecentRepos("broomva", 6),
-    getSafeSession({ fetchOptions: { headers: await headers() } }),
   ]);
 
   return (
@@ -39,13 +40,26 @@ export default async function Home() {
         type="application/ld+json"
         dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
       />
-      <LandingClient
-        projects={projects}
-        writing={writing}
-        notes={notes}
-        repos={repos}
-        userName={session.data?.user?.name}
-      />
+      <main className="relative">
+        {/* Hero personalization (session lookup) streams in via Suspense.
+            Anonymous fallback renders immediately, then swaps to the
+            personalized greeting when the session resolves. */}
+        <Suspense fallback={<HeroSection userName={null} />}>
+          <PersonalizedHero />
+        </Suspense>
+        <div className="mx-auto w-full max-w-6xl px-4 pb-24 sm:px-6">
+          <InstallSection />
+          <StackSection />
+          <ContentSection writing={writing} notes={notes} />
+        </div>
+      </main>
     </>
   );
+}
+
+async function PersonalizedHero() {
+  const session = await getSafeSession({
+    fetchOptions: { headers: await headers() },
+  });
+  return <HeroSection userName={session.data?.user?.name ?? null} />;
 }
