+++ apps/chat/components/site/landing-sections.tsx
@@ -5,11 +5,7 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useCallback, useMemo, useRef, useState } from "react";
 import { motion } from "motion/react";
-import type { ContentSummary } from "@/lib/content";
-import type { GitHubRepo } from "@/lib/github";
-import { ContentCard } from "@/components/site/content-card";
 import { ScrollReveal } from "@/components/site/scroll-reveal";
-import { formatDate } from "@/lib/date";
 import ThermodynamicGrid from "@/components/ui/interactive-thermodynamic-grid";
 
 /* ------------------------------------------------------------------ */
@@ -72,93 +68,11 @@ function PillIcon({ name }: { name: string }) {
   }
 }
 
-const stack = [
-  {
-    name: "bstack",
-    role: "Agent skills platform",
-    description:
-      "24 curated agent skills across 7 layers — install the full Broomva development workflow with one command.",
-    href: "/skills",
-  },
-  {
-    name: "Symphony",
-    role: "Orchestration",
-    description:
-      "Dispatch, lifecycle, and hook management for coding agents across issue trackers.",
-    href: "/projects/symphony",
-  },
-  {
-    name: "Control Kernel",
-    role: "Governance & safety",
-    description:
-      "Typed setpoints, safety shields, and multi-rate control loops that keep agents bounded.",
-    href: "/projects/control-metalayer",
-  },
-  {
-    name: "Life",
-    role: "Agent OS monorepo",
-    description:
-      "Arcan runtime, Lago persistence, Vigil observability, Praxis tool execution, Haima finance, and Spaces networking — unified in one Cargo workspace.",
-    href: "/projects/life",
-    demoHref: "/life",
-    demoLabel: "Try the live demo →",
-  },
-  {
-    name: "Autoany",
-    role: "Recursive improvement",
-    description:
-      "Evaluator-Governed Recursive Improvement (EGRI) — safe, measurable, rollback-capable optimization loops.",
-    href: "/projects/autoany",
-  },
-  {
-    name: "aiOS",
-    role: "Kernel contract",
-    description:
-      "The canonical type system — state vectors, event taxonomy, trait interfaces, and operating modes for all Agent OS crates.",
-    href: "/projects/aios",
-  },
-];
-
-/* ------------------------------------------------------------------ */
-/*  Props                                                              */
-/* ------------------------------------------------------------------ */
-
-interface LandingProps {
-  projects: ContentSummary[];
-  writing: ContentSummary[];
-  notes: ContentSummary[];
-  repos: GitHubRepo[];
-  userName?: string | null;
-}
-
-/* ------------------------------------------------------------------ */
-/*  Component                                                          */
-/* ------------------------------------------------------------------ */
-
-export function LandingClient({
-  projects,
-  writing,
-  notes,
-  repos,
-  userName,
-}: LandingProps) {
-  return (
-    <main className="relative">
-      <HeroSection userName={userName} />
-      <div className="mx-auto w-full max-w-6xl px-4 pb-24 sm:px-6">
-        <InstallSection />
-        <StackSection />
-        <ContentSection writing={writing} notes={notes} />
-      </div>
-    </main>
-  );
-}
-
 /* ------------------------------------------------------------------ */
 /*  Hero                                                               */
 /* ------------------------------------------------------------------ */
 
-function HeroSection({ userName }: { userName?: string | null }) {
+export function HeroSection({ userName }: { userName?: string | null }) {
   const router = useRouter();
   const [chatInput, setChatInput] = useState("");
   const inputRef = useRef<HTMLTextAreaElement>(null);
@@ -425,7 +339,7 @@ function HeroSection({ userName }: { userName?: string | null }) {
 /*  Install                                                            */
 /* ------------------------------------------------------------------ */
 
-function InstallSection() {
+export function InstallSection() {
   const [copied, setCopied] = useState(false);
   const cmd = "curl -fsSL https://broomva.tech/api/install | bash";
 
@@ -492,232 +406,3 @@ function InstallSection() {
     </section>
   );
 }
-
-/* ------------------------------------------------------------------ */
-/*  Stack                                                              */
-/* ------------------------------------------------------------------ */
-
-function StackSection() {
-  return (
-    <section className="mt-24 sm:mt-32">
-      <ScrollReveal>
-        <div className="glass rounded-3xl p-6 sm:p-10">
-          <p className="text-xs uppercase tracking-[0.25em] text-ai-blue">
-            Agent OS Stack
-          </p>
-          <h2 className="mt-3 font-display text-3xl text-text-primary sm:text-4xl">
-            The stack
-          </h2>
-          <p className="mt-3 max-w-2xl text-sm leading-relaxed text-text-secondary sm:text-base">
-            10 Rust crates, 24 agent skills, 500+ tests. A control-theory-native
-            ecosystem for autonomous software development.
-          </p>
-
-          <div className="mt-8 grid gap-4 sm:grid-cols-3">
-            {stack.map((item, i) => (
-              <ScrollReveal key={item.name} delay={i * 0.1} direction="left">
-                <div className="flex h-full flex-col gap-2">
-                  <Link
-                    href={item.href as Route}
-                    className="group glass-card block h-full transition hover:border-ai-blue/40"
-                  >
-                    <span className="font-mono text-[10px] uppercase tracking-[0.2em] text-ai-blue/50">
-                      {String(i + 1).padStart(2, "0")}
-                    </span>
-                    <p className="mt-2 font-display text-xl text-text-primary transition group-hover:text-ai-blue">
-                      {item.name}
-                    </p>
-                    <p className="mt-1 text-xs uppercase tracking-[0.15em] text-text-muted">
-                      {item.role}
-                    </p>
-                    <p className="mt-3 text-sm leading-relaxed text-text-secondary">
-                      {item.description}
-                    </p>
-                  </Link>
-                  {item.demoHref && (
-                    <Link
-                      href={item.demoHref as Route}
-                      className="inline-flex items-center gap-1.5 self-start rounded-full border border-ai-blue/25 bg-ai-blue/5 px-3 py-1.5 font-mono text-[10px] uppercase tracking-[0.18em] text-ai-blue/80 transition hover:border-ai-blue/60 hover:bg-ai-blue/10 hover:text-ai-blue"
-                    >
-                      {item.demoLabel ?? "Live demo →"}
-                    </Link>
-                  )}
-                </div>
-              </ScrollReveal>
-            ))}
-          </div>
-        </div>
-      </ScrollReveal>
-    </section>
-  );
-}
-
-/* ------------------------------------------------------------------ */
-/*  Recent Repos                                                       */
-/* ------------------------------------------------------------------ */
-
-function ReposSection({ repos }: { repos: GitHubRepo[] }) {
-  if (repos.length === 0) return null;
-
-  return (
-    <section className="mt-24 sm:mt-32">
-      <ScrollReveal>
-        <div className="glass rounded-3xl p-6 sm:p-10">
-          <p className="text-xs uppercase tracking-[0.25em] text-ai-blue">
-            Open Source
-          </p>
-          <h2 className="mt-3 font-display text-3xl text-text-primary sm:text-4xl">
-            Flagship Repos
-          </h2>
-          <div className="mt-8 grid gap-4 sm:grid-cols-2 lg:grid-cols-3">
-            {repos.map((repo, i) => (
-              <ScrollReveal key={repo.name} delay={i * 0.06}>
-                <a
-                  href={repo.html_url}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="group glass-card block h-full transition hover:border-ai-blue/40"
-                >
-                  <div className="flex items-start justify-between gap-2">
-                    <p className="font-display text-lg text-text-primary transition group-hover:text-ai-blue">
-                      {repo.name}
-                    </p>
-                    {repo.stargazers_count > 0 && (
-                      <span className="shrink-0 text-xs text-text-muted">
-                        &#9733; {repo.stargazers_count}
-                      </span>
-                    )}
-                  </div>
-                  {repo.description && (
-                    <p className="mt-1 line-clamp-2 text-sm text-text-muted">
-                      {repo.description}
-                    </p>
-                  )}
-                  {repo.language && (
-                    <p className="mt-2 text-xs text-text-muted/70">
-                      {repo.language}
-                    </p>
-                  )}
-                </a>
-              </ScrollReveal>
-            ))}
-          </div>
-        </div>
-      </ScrollReveal>
-    </section>
-  );
-}
-
-/* ------------------------------------------------------------------ */
-/*  Projects                                                           */
-/* ------------------------------------------------------------------ */
-
-function ProjectsSection({ projects }: { projects: ContentSummary[] }) {
-  return (
-    <section className="mt-24 sm:mt-32">
-      <ScrollReveal>
-        <div className="mb-6 flex items-end justify-between">
-          <h2 className="font-display text-3xl text-text-primary">
-            Pinned Projects
-          </h2>
-          <Link
-            href="/projects"
-            className="text-sm text-ai-blue transition hover:text-ai-blue/80"
-          >
-            View all
-          </Link>
-        </div>
-      </ScrollReveal>
-      <div className="grid gap-4 md:grid-cols-3">
-        {projects.map((project, i) => (
-          <ScrollReveal key={project.slug} delay={i * 0.08}>
-            <ContentCard
-              title={project.title}
-              summary={project.summary}
-              href={`/projects/${project.slug}`}
-              meta={formatDate(project.date)}
-              badge={project.status}
-            />
-          </ScrollReveal>
-        ))}
-      </div>
-    </section>
-  );
-}
-
-/* ------------------------------------------------------------------ */
-/*  Writing & Notes                                                    */
-/* ------------------------------------------------------------------ */
-
-function ContentSection({
-  writing,
-  notes,
-}: {
-  writing: ContentSummary[];
-  notes: ContentSummary[];
-}) {
-  return (
-    <section className="mt-24 grid gap-6 sm:mt-32 lg:grid-cols-2">
-      <div>
-        <ScrollReveal>
-          <div className="mb-6 flex items-end justify-between">
-            <h2 className="font-display text-3xl text-text-primary">
-              Latest Writing
-            </h2>
-            <Link
-              href="/writing"
-              className="text-sm text-ai-blue transition hover:text-ai-blue/80"
-            >
-              Read all
-            </Link>
-          </div>
-        </ScrollReveal>
-        <div className="grid gap-4">
-          {writing.map((entry, i) => (
-            <ScrollReveal key={entry.slug} delay={i * 0.08}>
-              <ContentCard
-                title={entry.title}
-                summary={entry.summary}
-                href={`/writing/${entry.slug}`}
-                meta={
-                  entry.readingTime
-                    ? `${formatDate(entry.date)} · ${entry.readingTime} min read`
-                    : formatDate(entry.date)
-                }
-              />
-            </ScrollReveal>
-          ))}
-        </div>
-      </div>
-
-      <div>
-        <ScrollReveal>
-          <div className="mb-6 flex items-end justify-between">
-            <h2 className="font-display text-3xl text-text-primary">
-              Recent Notes
-            </h2>
-            <Link
-              href="/notes"
-              className="text-sm text-ai-blue transition hover:text-ai-blue/80"
-            >
-              Browse notes
-            </Link>
-          </div>
-        </ScrollReveal>
-        <div className="grid gap-4">
-          {notes.map((entry, i) => (
-            <ScrollReveal key={entry.slug} delay={i * 0.08}>
-              <ContentCard
-                title={entry.title}
-                summary={entry.summary}
-                href={`/notes/${entry.slug}`}
-                meta={formatDate(entry.date)}
-              />
-            </ScrollReveal>
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-}
-
